Slight fixes

+++ internal/models/category.go
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS categories (
 	name                  VARCHAR(36) NOT NULL,
 	alias                 VARCHAR(128) NOT NULL,
 	description           VARCHAR(512) NOT NULL,
-	topics_count          INTEGER NOT NULL DEFAULT 0,
+	topics_count          BIGINT NOT NULL DEFAULT 0,
 	last_topic_id         VARCHAR(36),
 	position              INTEGER NOT NULL DEFAULT 0,
 	created_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
